revert some of the changes for gemini api until later
