+++ .github/workflows/knowledgeTesting.yml
@@ -1,4 +1,4 @@
-name: 1Cademy Testing
+name: Knowledge Testing
 on:
   pull_request:
     branches:
@@ -47,7 +47,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [18.x]
+        node-version: [16.x]
     steps:
       - uses: actions/checkout@v2
       - uses: isbang/compose-action@v1.3.2
@@ -57,10 +57,6 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
-      - name: Set Node.js environment variable
-        run: export NODE_OPTIONS=--openssl-legacy-provider
-      - name: Update caniuse-lite and browserslist
-        run: npm update caniuse-lite browserslist --legacy-peer-deps
       - name: npm install and build
         run: |
           npm install --legacy-peer-deps
