+++ package.json
@@ -33,7 +33,6 @@
     "@emotion/react": "11.9.3",
     "@emotion/server": "11.4.0",
     "@emotion/styled": "11.9.3",
-    "@google/generative-ai": "^0.13.0",
     "@mui/icons-material": "^5.8.4",
     "@mui/lab": "5.0.0-alpha.90",
     "@mui/material": "5.9.0",
